+++ src/ValueBuilders/AllMatchingTypesList.php
@@ -371,11 +371,16 @@ final class AllMatchingTypesList extends AllMatchingTypesListCache
             return self::fromClass($item);
         }
 
+        $reval = [];
+        // special case - is this a callable?
+        if (is_callable($item)) {
+            $retval[] = 'Callable';
+        }
+
         // if we get here, then this is just a plain old regular string
-        return [
-            "String",
-            self::FALLBACK_TYPE,
-        ];
+        $retval[] = "String";
+        $retval[] = self::FALLBACK_TYPE;
+        return $retval;
     }
 
     /**